+++ src/main/java/com/timcooki/jnuwiki/domain/docs/DTO/request/ContentEditReqDTO.java
@@ -1,4 +1,4 @@
-package com.timcooki.jnuwiki.domain.docs.DTO;
+package com.timcooki.jnuwiki.domain.docs.DTO.request;
 
 
 import lombok.Builder;
