+++ src/main/java/com/timcooki/jnuwiki/domain/docsRequest/service/DocsRequestService.java
@@ -1,7 +1,7 @@
 package com.timcooki.jnuwiki.domain.docsRequest.service;
 
-import com.timcooki.jnuwiki.domain.docs.DTO.InfoEditResDTO;
-import com.timcooki.jnuwiki.domain.docs.DTO.NewApproveResDTO;
+import com.timcooki.jnuwiki.domain.docs.DTO.response.InfoEditResDTO;
+import com.timcooki.jnuwiki.domain.docs.DTO.response.NewApproveResDTO;
 import com.timcooki.jnuwiki.domain.docs.entity.Docs;
 import com.timcooki.jnuwiki.domain.docsArchive.repository.DocsArchiveRepository;
 import com.timcooki.jnuwiki.domain.docsRequest.dto.request.NewWriteReqDTO;
