+++ src/main/java/com/timcooki/jnuwiki/domain/member/DTO/request/CheckNicknameReqDTO.java
@@ -1,4 +1,4 @@
-package com.timcooki.jnuwiki.domain.member.DTO;
+package com.timcooki.jnuwiki.domain.member.DTO.request;
 
 import lombok.Builder;
